+++ lib/mix/lib/mix/project.ex
@@ -343,18 +343,6 @@ defmodule Mix.Project do
       preferred_cli_env: %{"test" => :test} ]
   end
 
-  defp get_project_config(nil), do: []
-  defp get_project_config(atom) do
-    config = atom.project
-
-    if Keyword.has_key?(config, :env) do
-      IO.puts :stderr, "The :env key in #{inspect atom} project configuration is deprecated"
-    end
-
-    if env = config[:env][Mix.env] do
-      config |> Keyword.delete(:env) |> Keyword.merge(env)
-    else
-      config
-    end
-  end
+  defp get_project_config(nil),  do: []
+  defp get_project_config(atom), do: atom.project
 end